requirements.txt
streamlit==1.22.0
langchain==0.0.184
pinecone-client==2.2.1
openai==0.27.6
python-dotenv==1.0.0